refactor(jobs): bring system-jobs.ts under its line budget
check:architecture is failing on main. system-jobs.ts is 745 lines against a 700
budget, which arrived with PR #295. That matters more than a style nit because
verify.py aborts at its first failing phase and structural runs first, so one
over-budget file masks typecheck and tests for every branch in the repo — nobody
gets a deterministic verify signal until it is fixed.

The memory-dreaming outcome unit moves to a sibling, memory-dreaming-job-outcome.ts,
following the existing apps/core/src/jobs/* convention. system-jobs.ts drops to
652 with headroom rather than sitting at the cap.

Behaviour-neutral by construction and by proof: all six exports keep their
signatures, no existing test was modified, and the jobs suites pass 562 tests
across 53 files. scripts/ is untouched and no budget was raised — the owner chose
splitting over ratcheting, and D-0003 shows ratchets accumulate.

main needs BOTH this and PR #296 (which fixes group-processing.ts at 860/840) to
go green on this gate. check:architecture now reports only that file here.

Refs: docs/decisions/0055-arch1-client-signoff.md
Confirmed-by: Ravi
